Add API for specifying bulk stream ids
Being able to allocate bulk-streams is not really useful if the user cannot
specify the stream ids when submitting transfers.

Note that the actual stream id gets added to our private itransfer struct,
and a setter + getter is added to get to it. Unfortunately this is the only
way to add support for stream ids without breaking ABI. So this is another
item to fix when we do break ABI in libusb-2.0.

Most users will likely use the new libusb_fill_bulk_stream_transfer()
though, and will never know the difference.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>